+++ src/azure-cli/azure/cli/command_modules/appconfig/_validators.py
@@ -150,7 +150,20 @@ def validate_filter_parameter(string):
 
 
 def validate_assigned_identity(namespace):
-    return True
+    if namespace.assign_identity:
+        from msrestazure.tools import is_valid_resource_id
+        for identity in namespace.assign_identity:
+            if identity != '[system]' and not is_valid_resource_id(identity):
+                raise CLIError("Invalid identity '{}'. Use '[system]' to refer system assigned identity, or a resource id to refer user assigned identity.".format(identity))
+
+
+def validate_identities(namespace):
+    if namespace.identities:
+        from msrestazure.tools import is_valid_resource_id
+        for identity in namespace.identities:
+            if identity != '[system]' and identity != '[all]' and not is_valid_resource_id(identity):
+                raise CLIError("Invalid identity '{}'. Use '[system]' to refer system assigned identity '[all]' for all identities, or a resource id to refer user assigned identity.".format(identity))
+
 
 def validate_secret_identifier(namespace):
     """ Validate the format of keyvault reference secret identifier """