+++ src/azure-cli/azure/cli/command_modules/appconfig/custom.py
@@ -5,6 +5,8 @@
 
 # pylint: disable=line-too-long
 
+from knack.log import get_logger
+from knack.util import CLIError
 from azure.mgmt.appconfiguration.models import (ConfigurationStoreUpdateParameters,
                                                 ConfigurationStore,
                                                 Sku,
@@ -14,10 +16,18 @@ from azure.mgmt.appconfiguration.models import (ConfigurationStoreUpdateParamete
 from ._utils import resolve_resource_group, user_confirmation
 
 
+logger = get_logger(__name__)
+
+SYSTEM_ASSIGNED = 'SystemAssigned'
+USER_ASSIGNED = 'UserAssigned'
+SYSTEM_USER_ASSIGNED = 'SystemAssigned, UserAssigned'
+SYSTEM_ASSIGNED_IDENTITY = '[system]'
+
+
 def create_configstore(client, resource_group_name, name, location, sku, assign_identity=None):
-    configstore_params = ConfigurationStore(location.lower(),
+    configstore_params = ConfigurationStore(location=location.lower(),
                                             identity=__get_resource_identity(assign_identity) if assign_identity else None,
-                                            sku=Sku(sku))
+                                            sku=Sku(name=sku))
 
     return client.create(resource_group_name, name, configstore_params)
 
@@ -63,17 +73,95 @@ def configstore_update_set(cmd,
                          config_store_update_parameters=update_params)
 
 
-def configstore_update_custom(instance, tags=None, sku=None, assign_identity=None):
+def configstore_update_custom(instance, tags=None, sku=None):
     if tags is not None:
         instance.tags = tags
 
     if sku is not None:
-        instance.sku = Sku(sku)
-
-    if assign_identity is not None:
-        instance.assign_identity = __get_resource_identity(assign_identity)
+        instance.sku = Sku(name=sku)
 
     return instance
+
+
+def assign_identity(cmd, client, name, resource_group_name=None, identities=None):
+    if resource_group_name is None:
+        resource_group_name, _ = resolve_resource_group(cmd, name)
+
+    if identities is None:
+        identities = [SYSTEM_ASSIGNED_IDENTITY]
+
+    if '[all]' in identities:
+        raise CLIError("[all] is not supported for identity assignment")
+
+    current_identities = show_identity(cmd, client, name, resource_group_name)
+    user_assigned_identities = {}
+    identity_types = set()
+
+    if current_identities:
+        identity_types = identity_types if current_identities.type == 'None' else {identity_type.strip() for identity_type in current_identities.type.split(',')}
+        user_assigned_identities = current_identities.user_assigned_identities if current_identities.user_assigned_identities else {}
+
+    if SYSTEM_ASSIGNED_IDENTITY in identities:
+        identities.remove(SYSTEM_ASSIGNED_IDENTITY)
+        identity_types.add(SYSTEM_ASSIGNED)
+
+    user_assigned_identities.update({identity:UserIdentity() for identity in identities})
+    if user_assigned_identities:
+        identity_types.add(USER_ASSIGNED)
+
+    managed_identities = ResourceIdentity(type=','.join(identity_types) if identity_types else 'None',
+                                          user_assigned_identities=user_assigned_identities if user_assigned_identities else None)
+
+    client.update(resource_group_name=resource_group_name,
+                  config_store_name=name,
+                  config_store_update_parameters=ConfigurationStoreUpdateParameters(identity=managed_identities))
+
+    # Due to a bug in RP https://msazure.visualstudio.com/Azure%20AppConfig/_workitems/edit/6017040
+    # It client.update does not return the updated identities.
+    return show_identity(cmd, client, name, resource_group_name)
+
+
+def remove_identity(cmd, client, name, resource_group_name=None, identities=None):
+    if resource_group_name is None:
+        resource_group_name, _ = resolve_resource_group(cmd, name)
+
+    current_identities = show_identity(cmd, client, name, resource_group_name)
+    if not current_identities or current_identities.type == 'None':
+        logger.warning("No identity associated with this App Configuration.")
+        return
+
+    if not identities:
+        identities = [SYSTEM_ASSIGNED_IDENTITY]
+
+    user_assigned_identities = {}
+    if '[all]' in identities:
+        identity_types = None
+    else:
+        identity_types = {identity_type.strip() for identity_type in current_identities.type.split(',')}
+
+        if current_identities.user_assigned_identities:
+            for identity in current_identities.user_assigned_identities:
+                if identity not in identities:
+                    user_assigned_identities[identity] = current_identities.user_assigned_identities[identity]
+
+        if SYSTEM_ASSIGNED_IDENTITY in identities:
+            identity_types.discard(SYSTEM_ASSIGNED)
+
+        if not user_assigned_identities:
+            identity_types.discard(USER_ASSIGNED)
+
+    managed_identities = ResourceIdentity(type=','.join(identity_types) if identity_types else 'None',
+                                          user_assigned_identities=user_assigned_identities if user_assigned_identities else None)
+
+    client.update(resource_group_name=resource_group_name,
+                  config_store_name=name,
+                  config_store_update_parameters=ConfigurationStoreUpdateParameters(identity=managed_identities))
+
+
+def show_identity(cmd, client, name, resource_group_name=None):
+    config_store = show_configstore(cmd, client, name, resource_group_name)
+
+    return config_store.identity
 
 
 def list_credential(cmd, client, name, resource_group_name=None):
@@ -97,20 +185,20 @@ def __get_resource_identity(assign_identity):
     system_assigned = False
     user_assigned = {}
     for identity in assign_identity:
-        if identity == '[system]':
+        if identity == SYSTEM_ASSIGNED_IDENTITY:
             system_assigned = True
         else:
             user_assigned[identity] = UserIdentity()
 
     if system_assigned:
-        identity_type = "SystemAssigned"
+        identity_type = SYSTEM_ASSIGNED
     elif user_assigned:
-        identity_type = "UserAssigned"
+        identity_type = USER_ASSIGNED
     else:
         identity_type = "None"
 
     if system_assigned and user_assigned:
-        identity_type = "SystemAssigned, UserAssigned"
+        identity_type = SYSTEM_USER_ASSIGNED
 
     return ResourceIdentity(type=identity_type,
                             user_assigned_identities=user_assigned if user_assigned else None)